% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{add}
\usage{
add(..., na.rm = TRUE)
}
\arguments{
\item{...}{objects compatible with the `+` operator}

\item{na.rm}{passed to `Reduce()`}
}
\description{
Add all of the elements of a list together
}
