% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_same.R
\name{all_same}
\alias{all_same}
\title{all_same}
\usage{
all_same(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{passed to [`all_true()`]}
}
\description{
Test whether all elements of a vector are equal
}
\details{
Uses `==` to test for sameness.
}
