% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_true.R
\name{all_true}
\alias{all_true}
\title{all_true}
\usage{
all_true(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{passed to `all()`}
}
\description{
Test whether all elements of a vector are `TRUE`
}
\details{
Uses `all()` to test for true-ness.
}
