% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_all.R
\name{are_all}
\alias{are_all}
\title{Test whether two vectors are equal}
\usage{
are_all(e1, e2, na.rm = FALSE)
}
\arguments{
\item{e1}{vector}

\item{e2}{vector}

\item{na.rm}{whether to ignore \code{NA}s (default \code{FALSE})}
}
\value{
\code{TRUE} or \code{FALSE} (\code{NA}s will yield \code{FALSE} unless \code{na.rm = TRUE})
}
\description{
Test whether two vectors are equal
}
\examples{
\dontrun{
c("foo", "foo") \%>\% are_all("foo")
}
}
