% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkers.R
\name{chunk}
\alias{chunk}
\title{Chunk a list into multiple lists of a specific length.}
\usage{
chunk(xs, n)
}
\arguments{
\item{xs}{list. The list to chunk.}

\item{n}{numeric. The length of each chunk.}
}
\description{
Chunk a list into multiple lists of a specific length.
}
