% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mic.R
\name{clean_mic}
\alias{clean_mic}
\title{Clean MIC-values}
\usage{
clean_mic(df)
}
\arguments{
\item{df}{Data frame holding the MIC-values, rows = samples, columns = MIC values, first column named "id"}
}
\description{
Clean MIC-value data from SensiTitre
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
