% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mic_dist.R
\name{create_mic_dist}
\alias{create_mic_dist}
\title{Create MIC-distribution table}
\usage{
create_mic_dist(df, group_col)
}
\arguments{
\item{df}{The data frame holding the MIC-values and grouping data}

\item{group_col}{the name of the column with the grouping data}
}
\description{
Create a MIC-distribution table from clean MIC-data from SensiTitre
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
