% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{div}
\alias{div}
\title{Determine if a dividend is evenly divisible by a divisor.}
\usage{
div(x, y)
}
\arguments{
\item{x}{numeric. The dividend.}

\item{y}{numeric. The divisor.}
}
\description{
Determine if a dividend is evenly divisible by a divisor.
}

