% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compose.R
\name{fandcompose}
\alias{fandcompose}
\title{Compose two binary functions using and.}
\usage{
fandcompose(f1, f2)
}
\arguments{
\item{f1}{function. A function to compose.}

\item{f2}{function. A function to compose.}
}
\value{
TRUE or FALSE whether the results of both functions are TRUE.
}
\description{
Compose two binary functions using and.
}

