% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{filtermap}
\alias{filtermap}
\title{Filter a list to certain elements, apply a function to those elements, and return the initial list with the changed elements.}
\usage{
filtermap(xs, filter_f, map_f)
}
\arguments{
\item{xs}{list. The list to iterate over.}

\item{filter_f}{function. The predicate function to filter the list.}

\item{map_f}{function. The function to apply to each element of the filter.}
}
\description{
Filter a list to certain elements, apply a function to those elements, and return the initial list with the changed elements.
}

