% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{find}
\alias{find}
\title{Returns the first value of a list that meets a predicate.}
\usage{
find(xs, f)
}
\arguments{
\item{xs}{list. The list to search.}

\item{f}{function. The predicate function to apply. It should return TRUE or FALSE for each name element of the list.}
}
\description{
Returns the first value of a list that meets a predicate.
}

