% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{flatmap}
\alias{flatmap}
\title{Apply a map and then flatten the result.}
\usage{
flatmap(xs, f)
}
\arguments{
\item{xs}{list. The list to map over.}

\item{f}{function. The mapping function to apply to each element of the list.}
}
\description{
Apply a map and then flatten the result.
}

