% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{A shorthand for writing out R functions.}
\usage{
fn(...)
}
\arguments{
\item{...}{list. The last argument specifies the funciton body, the non-last
arguments specify the formals.}
}
\description{
A shorthand for writing out R functions.
}
\examples{
add <- fn(x, y, x + y)
add(2, 3)
}
