% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{forcompose}
\alias{forcompose}
\title{Compose two binary functions using or.}
\usage{
forcompose(f1, f2)
}
\arguments{
\item{f1}{function. A function to compose.}

\item{f2}{function. A function to compose.}
}
\value{
TRUE or FALSE whether the result of one or both functions is TRUE.
}
\description{
Compose two binary functions using or.
}

