% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_paste.R
\name{func_paste}
\alias{func_paste}
\title{Paste together unique values}
\usage{
func_paste(x, collapse = ",")
}
\arguments{
\item{x}{The column holding the vaues}

\item{collapse}{Delimiter between the unique values, default ","}
}
\description{
Paste together unique values in a column, and collapse with specified symbol.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
