% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%_>\%}
\alias{\%_>\%}
\title{Infix operator for reduce.}
\usage{
lhs \%_>\% rhs
}
\arguments{
\item{lhs}{list. The list to rduce.}

\item{rhs}{function. The reducing function.}
}
\description{
Infix operator for reduce.
}

