% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%_/>\%}
\alias{\%_/>\%}
\title{Infix operator for reducemap.}
\usage{
lhs \%_/>\% rhs
}
\arguments{
\item{lhs}{list. The list of lists to reduce.}

\item{rhs}{function. The function to reduce with.}
}
\description{
Infix operator for reducemap.
}

