% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%:>\%}
\alias{\%:>\%}
\title{Infix operator for filter.}
\usage{
lhs \%:>\% rhs
}
\arguments{
\item{lhs}{list. The list to filter.}

\item{rhs}{function. The function to filter with.}
}
\description{
Infix operator for filter.
}

