% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%:/>\%}
\alias{\%:/>\%}
\title{Infix operator for filtermap}
\usage{
lhs \%:/>\% rhs
}
\arguments{
\item{lhs}{list. The list to iterate over.}

\item{rhs}{list. A length-2 list of functions with teh first element being the filtering function and the second element being the mapping function.}
}
\description{
Infix operator for filtermap
}
