% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%===\%}
\alias{\%===\%}
\title{Determine whether two objects are identical.}
\usage{
x \%===\% y
}
\arguments{
\item{x}{object. The object to compare.}

\item{y}{object. The other object to compare.}
}
\description{
Determine whether two objects are identical.
}
