% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R, R/utils.R
\name{\%-=\%}
\alias{\%-=\%}
\alias{\%==\%}
\title{In-place subtraction.}
\usage{
x \%-=\% y

x \%==\% y
}
\arguments{
\item{x}{numeric. Minuend.}

\item{y}{numeric. Subtrahend.}

\item{x}{object. The object to compare.}

\item{y}{object. The other object to compare.}
}
\description{
In-place subtraction.

Determine whether two objects are identical.
}

