% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%f/>\%}
\alias{\%f/>\%}
\title{Infix operator for flatmap.}
\usage{
lhs \%f/>\% rhs
}
\arguments{
\item{lhs}{list. The list to map over.}

\item{rhs}{function. The function to map with.}
}
\description{
Infix operator for flatmap.
}

