% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/append.R
\name{\%<\%}
\alias{\%<\%}
\title{Infix operator for appending to a list.}
\usage{
xs \%<\% x
}
\arguments{
\item{xs}{list. The list to append to.}

\item{x}{object. The object to append to the list.}
}
\description{
Infix operator for appending to a list.
}

