% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{\%<<\%}
\alias{\%<<\%}
\title{Infix operator for in-place appending to list.}
\usage{
xs \%<<\% y
}
\arguments{
\item{xs}{list. The list to append to.}

\item{y}{object. The object to append to the list.}
}
\description{
Infix operator for in-place appending to list.
}
