% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%|0\%}
\alias{\%|0\%}
\title{Compose on length-0 values.}
\usage{
x \%|0\% y
}
\arguments{
\item{x}{object. This object will be returned if non-empty.}

\item{y}{object. This object will be returned if empty.}
}
\value{
either x or y.
}
\description{
Compose on length-0 values.
}

