% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%|all\%}
\alias{\%|all\%}
\title{Compose on NA, NULL, or lenth-0 (empty) values.}
\usage{
x \%|all\% y
}
\arguments{
\item{x}{object. This object will be returned if non-NA, non-NULL, non-empty.}

\item{y}{object. This object will be returned if x is NA, NULL, or empty.}
}
\value{
either x or y.
}
\description{
Compose on NA, NULL, or lenth-0 (empty) values.
}

