% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%|na\%}
\alias{\%|na\%}
\title{Compose on NA values.}
\usage{
x \%|na\% y
}
\arguments{
\item{x}{object. This object will be returned if non-NA.}

\item{y}{object. This object will be returned if x is NA.}
}
\value{
either x or y.
}
\description{
Compose on NA values.
}

