% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{\%|null\%}
\alias{\%|null\%}
\title{Compose on NULL values.}
\usage{
x \%|null\% y
}
\arguments{
\item{x}{object. This object will be returned if non-NULL.}

\item{y}{object. This object will be returned if x is NULL.}
}
\value{
either x or y.
}
\description{
Compose on NULL values.
}

