% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%/>\%}
\alias{\%/>\%}
\title{Infix operator for map.}
\usage{
lhs \%/>\% rhs
}
\arguments{
\item{lhs}{list. The list to map over.}

\item{rhs}{function. The mapping function}
}
\description{
Infix operator for map.
}

