% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%///>\%}
\alias{\%///>\%}
\title{Infix operator for innerinnermap.}
\usage{
lhs \%///>\% rhs
}
\arguments{
\item{lhs}{list. The list of lists top map.}

\item{rhs}{function. The function to map with.}
}
\description{
Infix operator for innerinnermap.
}

