% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign.R
\name{\%*=\%}
\alias{\%*=\%}
\title{In-place multiplication}
\usage{
x \%*=\% y
}
\arguments{
\item{x}{numeric. The multiplier.}

\item{y}{numeric. The multiplicand.}
}
\description{
In-place multiplication
}

