% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{\%v>\%}
\alias{\%v>\%}
\title{Infix operator for unlisting a list prior to mapping with a vectorized
function.}
\usage{
lhs \%v>\% rhs
}
\arguments{
\item{lhs}{list. The left-hand side list to unlist and apply.}

\item{rhs}{function. The right-hand side vectorized function to apply to the
unlisted vector.}
}
\description{
Infix operator for unlisting a list prior to mapping with a vectorized
function.
}
