% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepv}
\alias{grepv}
\title{Find all the matches of a pattern within a string.}
\usage{
grepv(haystack, needle, ...)
}
\arguments{
\item{haystack}{character. The string to search within.}

\item{needle}{character. The pattern to look for.}

\item{...}{list. Additional arguments to pass to grep.}
}
\description{
Find all the matches of a pattern within a string.
}
