% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{innerinnermap}
\alias{innerinnermap}
\title{Iterate a function over each subsublist in a list of list of lists. Recursion!}
\usage{
innerinnermap(xs, f, ...)
}
\arguments{
\item{xs}{list. The list-of-lists-of-lists to iterate over.}

\item{f}{function. The function to apply.}

\item{...}{list. Additional optional arguments to pass to lapply.}
}
\description{
Iterate a function over each subsublist in a list of list of lists. Recursion!
}

