% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{innermap}
\alias{\%//>\%}
\alias{innermap}
\title{Iterate a function over each sublist in a list of lists.}
\usage{
innermap(xs, f, ...)

lhs \%//>\% rhs
}
\arguments{
\item{xs}{list. The list-of-lists to iterate over.}

\item{f}{function. The function to apply.}
}
\description{
Iterate a function over each sublist in a list of lists.
}

