% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instPkgsPlusDeps.R
\name{instPkgPlusDeps}
\alias{instPkgPlusDeps}
\title{Install package and its dependencies}
\usage{
instPkgPlusDeps(
  pkg,
  install = FALSE,
  which = c("Depends", "Imports", "LinkingTo"),
  inc.pkg = TRUE
)
}
\arguments{
\item{pkg}{The package name}

\item{install}{If FALSE (default), the function returns a list of the dependencies. If TRUE, the function installs these dependencies}

\item{which}{The type of dependency, see \code{\link[tools]{package_dependencies}}.}

\item{inc.pkg}{If TRUE, installs the package in question in addition to the dependencies. If FALSE, only the dependencies are installed.}
}
\description{
This function is taken from the Stack Overflow issue https://stackoverflow.com/questions/21010705/update-a-specific-r-package-and-its-dependencies.
}
