% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is_lower}
\alias{is_lower}
\title{Determine whether a character is lowercase.}
\usage{
is_lower(x)
}
\arguments{
\item{x}{character. The character to look at.}
}
\description{
Determine whether a character is lowercase.
}

