% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_upper}
\alias{is_upper}
\title{Determine whether a character is uppercase.}
\usage{
is_upper(x)
}
\arguments{
\item{x}{character. The character to look at.}
}
\description{
Determine whether a character is uppercase.
}

