% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/joiners.R
\name{join}
\alias{join}
\title{Join a list of strings together into one string.}
\usage{
join(l, join = "")
}
\arguments{
\item{l}{list. The list to join.}

\item{join}{character. The string to join on.}
}
\description{
Join a list of strings together into one string.
}

