% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{lsplit}
\alias{lsplit}
\title{Split a string into a list by letters.}
\usage{
lsplit(str)
}
\arguments{
\item{str}{character. The string to split.}
}
\description{
Split a string into a list by letters.
}

