% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zippers.R
\name{lzip}
\alias{lzip}
\alias{\%+>\%}
\title{Zip two lists together into a list of lists containing each element from both lists.}
\usage{
lzip(xs, ys)

xs \%+>\% ys
}
\arguments{
\item{xs}{list. One list to zip.}

\item{ys}{list. Antoher list to zip.}
}
\description{
Zip two lists together into a list of lists containing each element from both lists.
}
