% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{map}
\alias{\%/>\%}
\alias{map}
\title{Iterate a function over a list.}
\usage{
map(xs, f, ...)

lhs \%/>\% rhs
}
\arguments{
\item{xs}{list. The list to iterate over.}

\item{f}{function. The function to apply.}
}
\description{
Iterate a function over a list.
}

