% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md5.R
\name{md5}
\alias{md5}
\title{Calculate the MD5 digest of an object}
\usage{
md5(x, ...)
}
\arguments{
\item{x}{any R object}

\item{\dots}{further arguments to \code{digest::digest}}
}
\value{
MD5 digest (a.k.a. hash) of \code{x}
}
\description{
Calculate the MD5 digest of an object
}
