% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.list}
\alias{merge.list}
\title{Merge two lists together.}
\usage{
\method{merge}{list}(xs, ys)
}
\arguments{
\item{xs}{list. A list to merge.}

\item{ys}{list. A list to merge.}
}
\description{
Merge two lists together.
}

