% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{nmap}
\alias{nmap}
\title{Iterate a function over the names of a list.}
\usage{
nmap(xs, f)
}
\arguments{
\item{xs}{list. The list to iterate over.}

\item{f}{function. The function to apply.}
}
\description{
Iterate a function over the names of a list.
}
