% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_calc.R
\name{perc_calc}
\alias{perc_calc}
\title{Calculate the percent occurrence of 1 and 0 in columns}
\usage{
perc_calc(df)
}
\arguments{
\item{df}{The data frame holding the values. Here, each column need to be one variable, f.ex. presence/absence of a gene, and the rows represent observations. Only values 1/0 are allowed in the data frame. Column names act as variable names in the resulting data frame}
}
\description{
The function calculates the percent occurrence of each column variable for all rows, where one row is one observation.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
