% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\title{pmean}
\usage{
pmean(..., na.rm = TRUE)
}
\arguments{
\item{...}{one or more vectors}

\item{na.rm}{passed to `mean()`}
}
\value{
double precision
}
\description{
Compute means across vectors (similar to `pmax` or `pmin`). Lazy evaluation meas this will work inside a `dplyr::mutate()`.
}
