% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preduce.R
\name{preduce}
\alias{preduce}
\title{Parallel reduce}
\usage{
preduce(.x, .f, ...)
}
\arguments{
\item{.x}{see [reduce]}

\item{.f}{see [reduce]}

\item{...}{see [reduce]}

\item{.init}{see [reduce]}

\item{dir}{see [reduce]}
}
\description{
Parallel reduce
}
\examples{
X <- mtcars[, c("cyl", "mpg")]
preduce(X, sum)

}
