% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressively.R
\name{progressively}
\alias{progressively}
\title{progressively}
\usage{
progressively(
  .f,
  total,
  title = "",
  format = "[:bar] :percent (:current/:total) elapsed: :elapsedfull eta: :eta",
  width = 80,
  show_after = 0,
  ...
)
}
\arguments{
\item{.f}{function}

\item{title}{(character) prepended to progress bar}

\item{format}{passed to [progress::progress_bar]}

\item{width}{passed to [progress::progress_bar]}

\item{...}{further arguments to [progress::progress_bar]}
}
\description{
Wrap a function in a progress bar.
Good when the function is called repeatedly, as with [purrr::map()].
}
