% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Import each excel sheet}
\usage{
read_excel_allsheets(filename, tibble = TRUE)
}
\arguments{
\item{filename}{The name of the excel file}

\item{tibble}{output as tibble or data frame? (TRUE/FALSE)}
}
\description{
Import each excel sheet and concatenate to a data frame or tibble.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@protonmail.com}
}
