% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{reduce}
\alias{reduce}
\title{Recursively recombine values of a list.}
\usage{
reduce(xs, f, init)
}
\arguments{
\item{xs}{list. The list to reduce.}

\item{f}{function. The reducing function to apply, taking two values and producing one value.}

\item{init}{object. An optional initialization value to start reducing.}
}
\description{
Recursively recombine values of a list.
}
