% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{reducemap}
\alias{reducemap}
\title{Reduce each sublist in a list of lists.}
\usage{
reducemap(xs, f)
}
\arguments{
\item{xs}{list. The list of lists to reduce.}

\item{f}{function. The reducing function to apply.}
}
\value{
a list where each element of the list of lists is reduced.
}
\description{
Reduce each sublist in a list of lists.
}

