% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of.R
\name{sum_of}
\alias{sum_of}
\title{Parallel sum, with tidyselect}
\usage{
sum_of(...)
}
\arguments{
\item{...}{objects that can be passed to [sum()]}
}
\description{
Parallel sum, with tidyselect
}
\examples{
library(dplyr)
mutate(mtcars, foo := sum_of(cyl, mpg))

}
