% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symdiff}
\alias{symdiff}
\title{Symmetric difference
Find all the elements that are not in both sets.}
\usage{
symdiff(xs, ys)
}
\arguments{
\item{xs}{list. The first list to look at.}

\item{ys}{list. The second element to look at.}
}
\description{
Symmetric difference
Find all the elements that are not in both sets.
}
