% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{thread}
\alias{thread}
\title{Thread a function through a list, iterating on an element and the previous
element.}
\usage{
thread(xs, fn)
}
\arguments{
\item{xs}{list. The list to iterate over.}

\item{fn}{function. A function that takes two arguments -- the previous
position and the current position -- and returns one element back.}
}
\description{
Thread a function through a list, iterating on an element and the previous
element.
}
