% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thread.R
\name{thread_while}
\alias{thread_while}
\title{Continue appending to a list while the functional is true, then start appending to the next list until the input list is empty.}
\usage{
thread_while(xs, fn)
}
\arguments{
\item{xs}{list. The input list to thread.}

\item{fn}{function. The function to check.}
}
\description{
Continue appending to a list while the functional is true, then start appending to the next list until the input list is empty.
}

