% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{vmap}
\alias{vmap}
\title{Iterate a function over the values of a list.}
\usage{
vmap(xs, f)
}
\arguments{
\item{xs}{list. The list to iterate over.}

\item{f}{function. The function to apply.}
}
\description{
Iterate a function over the values of a list.
}

