% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zippers.R
\name{vzip}
\alias{vzip}
\alias{zip}
\title{Zip two lists together into tuple vectors.}
\usage{
vzip(xs, ys)

zip(xs, ys)
}
\arguments{
\item{xs}{list. One list to zip.}

\item{ys}{list. Another list to zip.}
}
\description{
Zip two lists together into tuple vectors.
}
