% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitters.R
\name{wsplit}
\alias{wsplit}
\title{Split a string into a list by words.}
\usage{
wsplit(str)
}
\arguments{
\item{str}{character. The string to split.}
}
\description{
Split a string into a list by words.
}

