% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zippers.R
\name{zip_with}
\alias{zip_with}
\alias{\%+/>\%}
\title{Zip two lists together by a specific function.}
\usage{
zip_with(xs, ys, f)

xs \%+/>\% f
}
\arguments{
\item{xs}{list. One list to zip.}

\item{ys}{list. Antoher list to zip.}

\item{f}{function. A funciton that should take two arguments, iterating over each element of both lists.}
}
\description{
Zip two lists together by a specific function.
}
