% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{\%a-\%}
\alias{\%a-\%}
\title{In-place assignment, allowing for multiple values.}
\usage{
vars \%a-\% values
}
\arguments{
\item{vars}{character. The variable names to assign to. Can be a single
string or list of strings.}

\item{values}{object. The values to assign. Can be a single object or a
list of objects.}
}
\description{
In-place assignment, allowing for multiple values.
}
