% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functionals.R
\name{map}
\alias{map}
\title{Iterate a function over a list.}
\usage{
map(X, FUN, ...)
}
\arguments{
\item{X}{list. The list to iterate over.}

\item{FUN}{function. The function to apply.}

\item{...}{list. Additional optional arguments to pass to lapply.}
}
\description{
Iterate a function over a list.
}

