% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.list}
\alias{merge.list}
\alias{lmerge}
\title{Merge two lists together.}
\usage{
\method{merge}{list}(x, y, ...)

lmerge(x, y, ...)
}
\arguments{
\item{x}{list. A list to merge.}

\item{y}{list. A list to merge.}

\item{...}{list. Additional arguments that are ignored.}
}
\description{
Merge two lists together.
}
