% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkers.R
\name{partition}
\alias{partition}
\title{Paritition a list into two lists based on the reuslts of a binary function.}
\usage{
partition(xs, f)
}
\arguments{
\item{xs}{list. The list to partition.}

\item{f}{function. A function that returns TRUE or FALSE for each element into the list, determining which partition each element ends up in.}
}
\description{
Paritition a list into two lists based on the reuslts of a binary function.
}
